# Vectorized Array writing improves its speed