--- a/src/io.jl
+++ b/src/io.jl
@@ -99,9 +99,7 @@ function Base.write(io::IO, d::MRCData; compress = :none)
     T = datatype(h)
     data = parent(d)
     fswap = bswapfromh(h.machst)
-    for i in eachindex(data)
-        @inbounds sz += write(newio, fswap(T(data[i])))
-    end
+    write(newio, (fswap.(T.(data))))
     close(newio)
     return sz
 end