@sethaxen 2020-12-16
Thanks for the PR! I can confirm that I'm seeing roughly a 60-fold speed-up with this fix, which is quite surprising, since the old version did not copy the data, while this change does. I thought maybe there was a speed-up for `AbstractArray`s, but I tried `MappedArrays` and saw only a 10-fold speed-up.

How we should probably handle this is to allocate a large but reasonably sized buffer array of type `T` and copy to it slices of `fswap`ed and converted outputs, which are each written to the file. Then we get the best of both worlds: fewer and smaller allocations, with vectorized io. This will be critical for writing memmapped maps, where presumably the map would take up too much memory to load. Then a keyword argument could allow the user to change the buffer size.

Would you like to tackle this? If not, I can open an alternate PR.