@ehgus 2020-12-17
I come up with splitting a big array into 4kb(4096byte) vectors.

## why 4kb?
 The reason vectorized operation is more faster is that cpu offers the SIMD instructions. What I know is that the operation can handle 128~258bytes. So I guess julia will fully utilize vectorized instructions when its size is over 512bytes.
 
 Since mmap loads file with 4kb increments, which is default memory page size,  doing vectorized operation for each 4kb vectors seem fair enough. 

The following is the benchmark of `write` for both normal file and mmapped file (1.8Gb file)

> ## normal file
>   0.000011 seconds (7 allocations: 16.344 KiB)
>   0.000000 seconds
>   0.000088 seconds (138 allocations: 11.828 KiB)
>   0.000065 seconds (2 allocations: 32 bytes)
>   0.000001 seconds
>   0.000000 seconds
>   0.000000 seconds
>   4.080435 seconds (4.13 M allocations: 3.685 GiB, 11.04% gc time)
>  11.232843 seconds
> ---- done writing ----
> writing correctness: true 
> 
> ## mmapped file
>   0.000038 seconds (7 allocations: 16.344 KiB)
>   0.000000 seconds
>   0.000256 seconds (138 allocations: 11.828 KiB)
>   0.000059 seconds (2 allocations: 32 bytes)
>   0.000001 seconds
>   0.000000 seconds
>   0.000000 seconds
>   4.276065 seconds (4.13 M allocations: 3.685 GiB, 4.02% gc time)
>  10.974639 seconds
> ---- done writing ----
> writing correctness: true 

You see that allocation size become twice compared to simple vectorized operation. I believe it causes a slight performance drop for normal file writing, but still acceptable for me. I guess array to subarray operation makes a copy for preventing change of array size while writing.