@ehgus 2021-01-21
I haved benchmarked `write` performance depending its `unit_vsize`.

**workstation spec**
- Intel(R) Xeon(R) Gold 6138 CPU @ 2.00GHz
-  more than 100Gb RAM
- centos 7

**Laptop spec**
- 11th Gen Intel(R) Core(TM) i5-1135G7 @ 2.40GHz
- 8Gb RAM
- windows 10

<img width="446" alt="bench_workstation" src="https://user-images.githubusercontent.com/18534737/105321491-de10bf80-5c0a-11eb-80d0-2cf53579a70f.png">

<img width="446" alt="bench_laptop" src="https://user-images.githubusercontent.com/18534737/105321497-e10bb000-5c0a-11eb-814b-39e722bff0cc.png">

y axis is log scale of unit_vsize, in which 12 mean  4kb(2^12 bytes = 4kb ). Let's see the laptop case first. You can see a peak at 2^14bytes' `unit_vsize`. Except for this, the `write` speed increases as `unit_vsize` increase, and memory starvation may occur when its size over 1Mb. 
Combining the result, 4kb `unit_vsize` looks reasonable unit_vsize.